+++ ha_client.py
@@ -9,7 +9,7 @@ from typing import Any
 from urllib.error import HTTPError, URLError
 from urllib.request import Request, urlopen
 
-VERSION = "0.10.2"
+VERSION = "0.10.12"
 
 
 SUPPORTED_ACTIONS = {